requirements.txt
python-telegram-bot
yt-dlp
ffmpeg-python